backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
motor==3.7.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.2.0
pydantic==2.9.2
boto3==1.42.89
slowapi==0.1.5
python-multipart==0.0.7

gunicorn==21.2.0
